requirements.txt
streamlit
numpy
pandas
scikit-learn
opencv-python
tensorflow
matplotlib
Pillow
